+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+* CLI now can be compiled to and used as Windows executable.
+
 ## 4.5.0
 
 * feat: add static code diagnostics `avoid-nested-conditional-expressions`, `prefer-correct-identifier-length`, `prefer-correct-type-name`, `prefer-first`, `prefer-last`.
@@ -22,7 +26,6 @@
 * feat: introduce summary report.
 * chore: deprecate documentation in Github repo.
 * chore: restrict `analyzer` version to `>=2.4.0 <2.7.0`.
-* CLI now can be compiled to and used as Windows executable.
 
 ## 4.5.0-dev.1