+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 4.6.0
 
 * feat: CLI now can be compiled to and used as compiled executable.